Update services.js
fix(back): Fixing the problem that was blocking the mail if no document were attached

+++ api/services.js
@@ -53,30 +53,36 @@ export default async function handler(req, res) {
             emailContent += `<p><strong>${key.replace(/(<([^>]+)>)/ig, '')}:</strong> ${value.replace(/(<([^>]+)>)/ig, '')}</p>`;
         }
 
-        const cdcFile = files.cdc?.[0];
         let attachments = [];
-        if (cdcFile && cdcFile.originalFilename) {
-            const originalName = cdcFile.originalFilename.toLocaleLowerCase();
-            const isValidExtension = originalName.endsWith(".pdf")
-            const extensionMatches = originalName.match(/\.[A-z]{3}/g);
-            const hasMultipleExtensions = extensionMatches && extensionMatches.length == 1;
-            if (!(isValidExtension && hasMultipleExtensions)) {
-                try {
-                    fs.unlinkSync(cdcFile.filepath);
-                } catch (err) {
-                    console.error('Erreur suppression fichier non conforme :', err);
-                }
-                return res.status(400).json({
-                    success: false,
-                    message: 'Le type de fichier est invalide ou suspect. Seuls les fichiers PDF simples sont acceptés.',
-                });
+
+if (files.cdc) {
+    const cdcFile = Array.isArray(files.cdc) ? files.cdc[0] : files.cdc;
+
+    if (cdcFile && cdcFile.originalFilename) {
+        const originalName = cdcFile.originalFilename.toLowerCase();
+        const isValidExtension = originalName.endsWith(".pdf");
+        const extensionMatches = originalName.match(/\.[A-z]{3}/g);
+        const hasMultipleExtensions = extensionMatches && extensionMatches.length === 1;
+
+        if (!(isValidExtension && hasMultipleExtensions)) {
+            try {
+                fs.unlinkSync(cdcFile.filepath);
+            } catch (err) {
+                console.error('Erreur suppression fichier non conforme :', err);
             }
-            attachments.push({
-                filename: cdcFile.originalFilename,
-                content: fs.createReadStream(cdcFile.filepath),
-                contentType: cdcFile.mimetype,
+            return res.status(400).json({
+                success: false,
+                message: 'Le type de fichier est invalide ou suspect. Seuls les fichiers PDF simples sont acceptés.',
             });
         }
+
+        attachments.push({
+            filename: cdcFile.originalFilename,
+            content: fs.createReadStream(cdcFile.filepath),
+            contentType: cdcFile.mimetype,
+        });
+    }
+}
 
         await transporter.sendMail({
             from: `"Site EPI Studios" <${process.env.EMAIL_FROM}>`,
